Have Bevy find assigned canvas element
Bevy creates it's own canvas element by default, but the custom page
defines its own. This makes Bevy go find that one.

The window sizing still works -- it simply resizes the canvas upon
starting. I'm not sure if I'll change that or remove the size from the
HTML. That's a tomorrow problem.